fix(report): read replay preferences from query
