+++ apps/report/src/components/detail-panel/index.tsx
@@ -34,13 +34,7 @@ const VIEW_TYPE_BLACKBOARD = 'blackboard';
 const VIEW_TYPE_SCREENSHOT = 'screenshot';
 const VIEW_TYPE_JSON = 'json';
 
-const DetailPanel = ({
-  focusOnCursor,
-  showElementMarkers,
-}: {
-  focusOnCursor?: boolean;
-  showElementMarkers?: boolean;
-}): JSX.Element => {
+const DetailPanel = (): JSX.Element => {
   const insightDump = useExecutionDump((store) => store.insightDump);
   const _contextLoadId = useExecutionDump((store) => store._contextLoadId);
   const activeExecution = useExecutionDump((store) => store.activeExecution);
@@ -90,8 +84,6 @@ const DetailPanel = ({
         replayScripts={animationScripts || []}
         imageWidth={imageWidth || 0}
         imageHeight={imageHeight || 0}
-        autoZoom={focusOnCursor}
-        elementsVisible={showElementMarkers}
       />
     );
   } else if (!activeTask) {
