+++ apps/report/src/types.ts
@@ -17,6 +17,4 @@ export interface PlaywrightTasks {
 export interface VisualizerProps {
   logoAction?: () => void;
   dumps?: PlaywrightTasks[];
-  focusOnCursor?: boolean;
-  showElementMarkers?: boolean;
 }
